Add Catalyst Center 3.2.3.0 support and address ansible-inclusion feedback
- 427 new action/module pairs from the 3.2.3.0 API, a dynamic inventory
  plugin, and AAP indirect node counting
- Corrected version_added across all modules and fixed required_if specs
  blocking valid create calls in 28 action plugins
- Fixed a doc bug that broke ansible-doc collection-wide and a broken
  exists() check in sites_telemetry_settings
- Restored ansible-core >=2.16 support to match Red Hat certification;
  updated certification workflow and added dependabot

+++ plugins/action/sda_port_assignments_info.py
@@ -33,7 +33,7 @@ argument_spec.update(
         interfaceName=dict(type="str"),
         dataVlanName=dict(type="str"),
         voiceVlanName=dict(type="str"),
-        nativeVlanId=dict(type="float"),
+        nativeVlanId=dict(type="int"),
         offset=dict(type="int"),
         limit=dict(type="int"),
         headers=dict(type="dict"),
